Task 3.5 course view

See merge request foxstudent102959/university-cms!5

+++ src/main/java/ua/com/foxminded/volodymyrtolpiekin/universitycms/service/impl/CourseServiceImpl.java
@@ -1,20 +1,32 @@
 package ua.com.foxminded.volodymyrtolpiekin.universitycms.service.impl;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
+import ua.com.foxminded.volodymyrtolpiekin.universitycms.dto.CourseDTO;
 import ua.com.foxminded.volodymyrtolpiekin.universitycms.models.Course;
 import ua.com.foxminded.volodymyrtolpiekin.universitycms.repository.CourseRepository;
+import ua.com.foxminded.volodymyrtolpiekin.universitycms.service.GroupService;
 import ua.com.foxminded.volodymyrtolpiekin.universitycms.service.CourseService;
+import ua.com.foxminded.volodymyrtolpiekin.universitycms.service.TutorService;
 
 import java.util.List;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 public class CourseServiceImpl implements CourseService {
     private final CourseRepository courseRepository;
+    private final GroupService groupService;
+    private final TutorService tutorService;
+    private final ModelMapper mapper;
 
-    public CourseServiceImpl(CourseRepository courseRepository) {
+    public CourseServiceImpl(CourseRepository courseRepository, GroupService groupService, TutorService tutorService, ModelMapper mapper) {
         this.courseRepository = courseRepository;
+        this.groupService = groupService;
+        this.tutorService = tutorService;
+        this.mapper = mapper;
     }
 
     @Override
@@ -40,7 +52,28 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public void deleteById(Long id){
-        findById(id);
-        courseRepository.deleteById(id);
+        if (courseRepository.existsById(id)) {
+            courseRepository.deleteById(id);
+        }
+    }
+
+    @Override
+    public List<CourseDTO> readAllDTOs() {
+        return findAll()
+                .stream()
+                .map(course -> mapper.map(course, CourseDTO.class))
+                .collect(toList());
+    }
+
+    @Override
+    public CourseDTO createCourse(CourseDTO courseDTO) {
+        Course course = mapper.map(courseDTO, Course.class);
+        addCourse(course);
+        return courseDTO;
+    }
+
+    @Override
+    public CourseDTO findDTOById(Long courseId) {
+        return mapper.map(findById(courseId), CourseDTO.class);
     }
 }
